requirements.txt
aiogram==2.19
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.12
frozenlist==1.3.0
idna==3.3
lxml==4.8.0
multidict==6.0.2
Pillow==9.1.0
python-pptx==0.6.21
pytz==2022.1
XlsxWriter==3.0.3
yarl==1.7.2
